requirements.txt
pillow
mutagen